refactor: remove redundant scripts
