--- workspace/python/ds-pack/2.1.0-base-python-3.8.18/vscode-tunnel/scripts/install.sh
@@ -1,1 +1,0 @@
-code tunnel
